fix: remote task dependency does not work
